+++ Sources/StreamChatSwiftUI/ViewFactory/DefaultViewFactory.swift
@@ -478,7 +478,7 @@ extension ViewFactory {
             messageController: options.messageController,
             quotedMessage: options.quotedMessage,
             editedMessage: options.editedMessage,
-            onMessageSent: options.onMessageSent
+            willSendMessage: options.willSendMessage
         )
     }